+++ persistence/src/interest.rs
@@ -1,4 +1,5 @@
 use async_trait::async_trait;
+use bigdecimal::BigDecimal;
 use sqlx::PgPool;
 use uuid::Uuid;
 
@@ -9,10 +10,10 @@ use business::interest::repository::InterestRepository;
 #[derive(Debug, sqlx::FromRow)]
 pub struct PersistenceInterestConfiguration {
     pub id: Uuid,
-    pub start_range: i64,
-    pub end_range: i64,
-    pub fine: f64,
-    pub interest: f64,
+    pub start_range: i32,
+    pub end_range: i32,
+    pub fine: BigDecimal,
+    pub interest: BigDecimal,
 }
 
 impl From<InterestConfiguration> for PersistenceInterestConfiguration {
